@@ -55,8 +55,8 @@
     "@types/enzyme": "^3.1.15",
     "@types/enzyme-adapter-react-16": "^1.0.3",
     "@types/jest": "^24.0.9",
-    "@types/react": "^16.8.23",
-    "@types/react-dom": "^16.8.5",
+    "@types/react": "^16.9.5",
+    "@types/react-dom": "^16.9.1",
     "babel-core": "^7.0.0-bridge.0",
     "cpy-cli": "^2.0.0",
     "cross-env": "^5.2.0",
@@ -66,8 +66,8 @@
     "enzyme-to-json": "^3.4.0",
     "handsontable": "^7.1.1",
     "jest": "^24.1.0",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
+    "react": "^16.10.2",
+    "react-dom": "^16.10.2",
     "react-redux": "^7.1.1",
     "redux": "^4.0.4",
     "rollup": "^1.17.0",
