@@ -62,6 +62,7 @@ describe('Using hooks within HotTable renderers', () => {
     simulateMouseEvent(hotInstance.getCell(0,0).querySelector('button'), 'click');
 
     expect(hotInstance.getCell(0,0).querySelector('span').innerHTML).toEqual('3');
+    expect(hotInstance.getCell(1,1).querySelector('span').innerHTML).toEqual('0');
 
     wrapper.detach();
     done();
@@ -69,7 +70,5 @@ describe('Using hooks within HotTable renderers', () => {
 });
 
 /*
-
  Editor components cannot be used with React Hooks, as they need to be classes derived from BaseEditorComponent.
-
  */
